Convert netlify function to typescript
