+++ .babelrc
@@ -1,0 +1,11 @@
+{
+  "presets": [
+    "@babel/preset-typescript",
+    [
+      "@babel/preset-env",
+      {
+        "targets": "node 16.0"
+      }
+    ]
+  ]
+}
